# docs/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: msmariaulfa33@gmail.com
  version: 0.0.1
  title: DNR Service Swagger Docs
  description: Service for DNR Website
servers:
  - description: Local development
    url: 'http://localhost:3001'
...
paths:
...
  /api/v1/orders:
    post:
      summary: Create Orders
      operationId: createOrder
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      description: Endpoint to create an order
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order-create'
      tags:
        - order
    parameters: []
    get:
      summary: Get Orders For User
      operationId: getOrdersForUser
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order-summary'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      tags:
        - order
      description: Endpoint to get list orders for the login user
      parameters:
        - schema:
            type: string
            enum:
              - ORDERED
              - PAID
              - CANCELED
              - EXPIRED
              - PROCESSED
              - COMPLETED
              - DELIVERED
              - ONGOING
          in: query
          name: status
  '/api/v1/orders/{orderId}':
    parameters:
      - schema:
          type: string
        name: orderId
        in: path
        required: true
    get:
      summary: Get Order By Id
      operationId: getOrderById
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      tags:
        - order
      description: Endpoint to get order with matching id
      parameters: []
    patch:
      summary: Update Order By Id
      operationId: updateOrderById
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      tags:
        - order
      description: Endpoint to update an order
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order-update'
  '/api/v1/orders/{orderId}/charge':
    parameters:
      - schema:
          type: string
        name: orderId
        in: path
        required: true
    post:
      summary: Charge Card Payment
      operationId: chargeCardPayment
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Charge-payment-success'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Card-charge-payment'
      description: Endpoint to charge payment card
      tags:
        - order
  '/api/v1/orders/{orderId}/cancel':
    post:
      summary: Cancel Orders
      operationId: cancelOrder
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      description: Endpoint to cancel an order
      tags:
        - order
    parameters:
      - schema:
          type: string
        name: orderId
        in: path
        required: true
  '/api/v1/orders/{orderId}/complete':
    post:
      summary: Complete Orders
      operationId: completeOrder
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      description: Endpoint to complete an order
      tags:
        - order
    parameters:
      - schema:
          type: string
        name: orderId
        in: path
        required: true
  /api/v1/admin/orders:
    parameters: []
    get:
      summary: Get Orders For Admin
      operationId: getOrdersForAdmin
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                  totalPage:
                    type: integer
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order-summary'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      tags:
        - admin
      description: Admin endpoint to get orders
      parameters:
        - schema:
            type: string
            enum:
              - ORDERED
              - CANCELED
              - PROCESSED
              - COMPLETED
              - DELIVERED
              - ONGOING
          in: query
          name: status
        - schema:
            type: string
          in: query
          name: page
          description: Get product page use for pagination
        - schema:
            type: string
          in: query
          name: limit
          description: Limit the return product
        - schema:
            type: string
          in: query
          name: search
          description: Limit the return product
  /api/v1/admin/orders/{id}:
    get:
      summary: Get Orders For Admin
      operationId: getOrdersForAdmin
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                  totalPage:
                    type: integer
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order-summary'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      tags:
        - admin
      description: Admin endpoint to get orders
    patch:
      summary: input invoice number by payment Id
      operationId: inputInvoiceNumber
      security:
        - ApiKeyAuth: []
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment-invoice'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
  /api/v1/admin/orders/{transactionNumber}/update-cart:
    patch:
      summary: Update Product on Invoice
      operationId: updateCartById
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      description: Endpoint to update cart location or quantity
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cart-invoice-update'
      tags:
        - cart
  /api/v1/admin/orders/{transactionNumber}:
    parameters:
      - schema:
          type: string
        name: transactionNumber
        in: path
        required: true
    get:
      summary: Get Order By Id
      operationId: getOrderByIdAdmin
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      tags:
        - order
      description: Endpoint to get order with matching id
    patch:
      summary: Update Product on Invoice
      operationId: updateCartById
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      description: Endpoint to update cart location or quantity
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cart-invoice-update'
      tags:
        - cart
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-auth-token
  schemas:
...
    Order:
      title: Order
      type: object
      description: Model representing order
      x-examples:
        - id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
          transaction_number: '123456.89012'
          user_id: a169451c-8525-4352-b8ca-070dd449a1a5
          shipment_id: e415c869-52a9-4dbd-bb4b-c8a6b3968a51
          payment_id: d43b87f9-9e28-4802-8eaa-6ee91a40ea71
          status: ORDERED
          carts:
            - id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
              product_id: 0d012afa-f885-4e65-aeca-37e27701e2d1
              user_id: a169451c-8525-4352-b8ca-070dd449a1a5
              location: Bekasi
              quantity: 5
              status: ACTIVE
              final_unit_price: 1000000
              order_id: 93101167-9065-4b9c-b98b-5d789a3ed9fe
              created: '2019-08-24T14:15:22Z'
              updated: '2019-08-24T14:15:22Z'
          expired_at: '2019-08-24T14:15:22Z'
          created: '2019-08-24T14:15:22Z'
          updated: '2019-08-24T14:15:22Z'
      properties:
        id:
          type: string
          format: uuid
        transaction_number:
          type: string
        order_events:
          type: array
          items:
            type: object
          nullable: true
        user_id:
          type: string
          format: uuid
        shipment:
          $ref: '#/components/schemas/Shipment'
        payment:
          $ref: '#/components/schemas/Payment'
        carts:
          type: array
          items:
            $ref: '#/components/schemas/Cart-summary'
        status:
          type: string
          enum:
            - ORDERED
            - PROCESSED
            - DELIVERED
            - COMPLETED
            - CANCELED
            - ONGOING
        expired_at:
          type: string
          format: date-time
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
      required:
        - id
        - transaction_number
        - user_id
        - shipment
        - payment
        - status
...